Refactor backend functions
